add: Add new estructure in order to not duplicate code
