+++ deps/wazuh_testing/wazuh_testing/fim_module/fim_variables.py
@@ -1,0 +1,1 @@
+max_events = 20
